Move coord transformation methods and fix use of lon_ lat_
